Update pipeline for new repo
Now that gpupgrade no longer needs to pull in the GPDB source, we can revert a
lot of the changes made to the original multi-cluster testing pipeline pulled
from https://github.com/greenplum-db/gp-performance-testing/tree/dev/pg_upgrade.

Specifically, this commit updates the pipeline file to do the following:
- Update task fields for the AWS -> GCP migration
- Remove the compile job
- Pull down a 6.X binary to replace the GPDB 6.X source
- Pull out the inline Bash setup code into a setup_clusters.sh script

The pipeline still doesn't run any tests once the clusters are set up; the
intention is for this to be a platform for manual testing, and automated tests
will be added to the pipeline later.

Authored-by: Jamie McAtamney <jmcatamney@pivotal.io>