Refactor ClusterPair
This commit refactors ClusterPair to be a pair of cluster.Cluster objects
instead of holding all of the segment information itself, to allow us to
leverage the functionality in the cluster package.

This has several consequences:
- Segment configuration information is now retrieved once and stored in memory,
  rather than being read from the config file every time.
- The configutils package has been removed and the Reader functionality has
  been added to ClusterPair.
- Quite a few testing-related structs are no longer necessary and have also been
  removed.
- Several tests have been modified to use the cluster's Executor for shelling
  out to Bash instead of CommandExecer, due to its simpler and more flexible
  interface (and CommandExecer will be removed entirely in a later commit).
- Helper functions for creating test Clusters and test ClusterPairs have been
  added to testutils to simplify test setup.

This commit breaks several tests and leaves them marked as Pending, because the
refactor uncovered flakiness and invalid behavior that would require too
extensive a refactor to fix, and this commit is already fairly large.  These
tests will be addressed in later refactor commits.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>