Add MockChecklistManager
This commit pulls the ChecklistManager mocking out of cluster_ssher into
testutils, makes it public to be usable in all tests, and refactors it for
efficiency and usability.  The integration tests checking for a process flow
of PENDING -> RUNNING -> COMPLETE will be refactored to use it in a later
commit.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Kevin Yeap <kyeap@pivotal.io>