Refactor tests to use MockChecklistManager
Previously, our tests were checking the status of individual steps by
calling out to status upgrade in the CLI, which required a lot of
mocking and made some tests flaky.  This commit refactors the tests to
use a MockChecklistManager, which simplifies the testing and reduces the
test time.

Additionally, the hub was refactored to take a ChecklistWriter
separately from the one in its RemoteExecutor, and several functions
were moved from the ClusterPair object to the Hub, in preparation for a
later refactor of ClusterSsher.

Finally, checks were added to the status upgrade step to not attempt to
gather status information for steps that could not have been run yet if
check config or prepare start-agents were not yet run, and the tests
were modified accordingly.

Co-authored-by: Kevin Yeap <kyeap@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>