Add RunCommandAsync function for running pg_upgrade
The cluster.Executor interface has no function for running asynchronous
commands, and the gpupgrade use case is not general enough to add such a
command to gp-common-go-libs, so this commit adds a RunCommandAsync
locally.

RunCommandAsync starts a shell command and redirects all output to a
specified log file.  It is assigned to a function pointer in utils.System
for mocking purposes.

Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>