Unify unit test initialization
Previously, every unit test file performed its own setup and teardown in
BeforeEach and AfterEach, most of which was just copy-pasted between test files,
so any refactor touching the hub or agent would generally involve making the
same change in lots of files.

In a similar manner to a previous commit[1], this one moves all the duplicated
logic to the BeforeEach and AfterEach in services_suite_test.go, so that the
initialization happens only once and refactors only need to make a given change
once.

[1] 01fba51a72708669f5b2afd8e275b933714edeef

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>