Execute: stream output for single-host steps
Stream stdout and stderr for execute sub-steps that don't need to stream
from multiple hosts.

InitTargetCluster had to be refactored to ignore warnings from
gpinitsystem instead of failing, because those warnings result in an
exit code of 1. This caused gpupgrade to incorrectly fail. This special
logic will be removed when gpinitsystem is fixed.

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>