Run pg_upgrade --check during initialize
Add a substep under the initialize step to run pg_upgrade --check against the
master and primaries and log the results in the state directory.  Both master
and primaries checks are done in a single substep.

Because pg_upgrade --check requires that the target cluster be initialized for
the check to work, this commit moves all cluster creation logic from execute to
initialize as well.  The diskSpace check is performed before the new cluster is
initialized, and the version check is disabled since the source cluster is
stopped where it would have made sense to do it.

We use "stop", "start", "old" and "new" in the UI output for steps to provide
uniform output for a user.

BATS and unit tests were added.  To make the negative tests fail, we add a
partition table with indexes in the source cluster.  We assume this will always
cause pg_upgrade to fail, but the tests will fail if pg_upgrade ever becomes
able to handle that case.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Mark Sliva <msliva@pivotal.io>
Co-authored-by: David Krieger <dkrieger@pivotal.io>