Initialize: Make creating directories and cluster configs idempotent
We split the current single function that both creates the state
directory and the initial cluster configs into two functions.  We then
make each of these steps idempotent by succeeding fast if
the directories already exist. This is in preparation for skipping
completed steps and continuing when an error occured.

Co-authored-by: David Krieger <dkrieger@pivotal.io>