hub accepts a --state-directory flag
For a given cluster upgrade, the CLI and Hub must agree on the
state directory used as the CLI and Hub are both required to run
on the master host.  Currently, we also want to Hub and Agent
state directory to match, even though the segment hosts need not
coincide with the master host.

To keep this consistent, we have the CLI pass its state directory
to the Hub via this new command flag.  The Hub alread does the same
thing with each agent.