Update help text for gpupgrade/initalize/execute/finalize
We hard-coded the help output for flags:
  "-?"
  "-h"
  "--help"
  "help"
based on the provided specification.

We added tests to make sure the flags output the required values.

For gpupgrade initialize, we changed persistent flags to local flags
to be able to implement the "-?" flag.  Since we do not have any
child commands of gpupgrade initialize that require these, this should
be fine.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>