Make help text dynamic
To help prevent a divergence in the descriptions of each substep printed during
execution and the descriptions printed as part of the help text, this commit
changes the help text for all steps to be dynamically generated.

The substep descriptions now include both kinds of help text in one place
(identical except for conjugation and punctuation) and the help text
interpolates the appropriate text using a list of substeps ordered as they
should be shown in the help.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>