fault injection
We add a basic fault injector.  To use it:

gpupgrade initialize --fault-injection ....

and then in any subsequent hub code:

if fault_injector.Insert("callMe", 10) {
    fmt.Println("callMe executing arbitrary code")
}

NOTE: this currently only runs fault injection code on the cli once
gpugprade initialize has been called and on the hub during and after
the call to gpugrade initialize starts the hub.