CI: stabilize install-tests
Specify a temp-port-range from 6020-6040 for all bats tests which is closer to
the working 6000-6007 range of the source cluster.  We have not seen any port
contention in the 6000-6007 range that the source cluster uses, so we stay near
that range for the target cluster to fix the ephemeral port conflicts we have
seen around 10-20% of the time.

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>