Silence linter errors for false positives.
The linter is detecting that the dialer returned from
mock_agent.NewMockAgentServer() is not of type hub.Dialer in
hub/services_suite_test.go: and hub/server_test.go. These appear to be
the same type, so silence. We are also planning on migrating these tests
from ginkgo to go-style.

Co-authored-by: Mark Sliva <msliva@pivotal.io>