proto: small refactor of hub-agent rename rpc command
We establish idempotence for the finalize substep FINALIZE_UPDATE_DATA_DIRECTORIES.

We use the atomicity of os.Rename() to make the actual UPDATE step
idempotent. This is detailed and tricky, as we must also handle idempotence
of the deletion of the mirrors/standby of the source cluster(in link mode)
as well as idempotence on the agents.

In addition, we use a new archive directory naming scheme to allow
a further cross-check on renaming.  Instead of the current scheme,
           DIR_PATH/sourceDir -->  DIR_PATH/sourceDir_old
we embed the ugprade ID:
           DIR_PATH/sourceDir --> DIR_PATH/sourceDir.ABC123.old