ci: filter retail-demo data dump
To transform the retail demo 6 dump to look like a 5 dump, a replacement
operation for regex substitutions is added. This is because dumping the
retail data before upgrade does not add single quotes around partition
table rel options. Fixing would require changes to GPDB 5.

This also adds a retail_demo.diff file which is patched to the data dump
after upgrade in order to compare the dumps. This is needed since the
dump before upgrade omits the default sequence START WITH clause. This
clause is dumped in GPDB 6 and upstream postgres.

Co-authored-by: David Krieger <dkrieger@pivotal.io>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>