rename master data directories idempotently
This adds upgrade.RenameDataDirectory to idempotently rename source to
archive, and target to source. This is placed in the upgrade package
since the following commit will use it to rename the segment data
directories as well. Thus, RenameDataDirectory is used in both the hub
for renaming the master data directories, and will be used in the agent
for renaming the segments.

To enable idempotence we allow ENOENT errors when renaming the source to
archive as this may occur when renaming the target fails resulting in
the source directory not existing.

Co-authored-by: David Krieger <dkrieger@pivotal.io>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>