rename segment data directories idempotently
The agents now call upgrade.RenameDataDirectory, which is also called by
the hub when renaming the master data directories.

In link mode the mirrors have been deleted to save disk space. Since the
upgraded mirrors will be added later to the correct directory there is
no need to rename target to source. Thus, the RenameTarget boolean is
used to indicate whether to rename the source to target.

Co-authored-by: David Krieger <dkrieger@pivotal.io>
Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>