use upgradeID when archiving the source directory
To mitigate naming collisions when archiving the source directory we use
the upgradeID in the archive name. For example, a source directory of
/data/dbfast1/demoDataDir0 is archived as
/data/dbfast1/demoDataDir.123ABC.0.old.

ArchiveAndSwapDirectories no longer takes the archive name as that is
determined internally within the function. This prevents the need for
callers to know the archive name, and simplifies logic. The archive name
is found through a shortcut by appending the old suffix to the target
instead of manipulating the source directly.