fix agent test for default port
Don't specify --port when testing the default.

By switching to cmd.CombinedOutput() and daemonizing the agent we
correctly bubble up errors if the agent fails to start. This also
drastically speeds up test time from 1 to 0.04 seconds. In order to
correctly kill the daemonized agent we need to enable process groups and
kill all processes in the group. We do so by sending a kill with -PGID
which in this case is the same as -PID.