generate SQL scripts that drop non-upgradable objects
Iniitialize runs a "pg_upgrade --check", which can find various
database objects that cannot be upgraded.  These scripts generate
SQL scripts that can be used to drop different types of these
objects.

The scripts do not drop the objects directly, as a user might want
to inspect the objects first.