kill-services: try hub default port if configFile cannot be accessed
kill-services is used internally for testing only.  Typically, a bats
test will call kill-services in both the setup() and teardown().
However, in setup() where kill-services is called, the configFile for
the current run is not set.  But if the hub is already running(say,
from a previous failed test or the developer has been running it),
kill-services will fail.  It should pass.

This fixes this case.  It is typically not currently seen because
kill-services exits quickly if the hub is not running.  And the
previous test's teardown() stops the hub.