implement execute link-mode revert
During revert, recover the original source cluster by rsync'ing the
untouched source cluster's standby and mirrors to the master and
primaries.  This has the effect of reverting any changes to the
original source cluster that may have occurred as a result of running
gpupgrade execute or via actions of the user for testing the target
cluster.

We also perform this rsync for copy mode.  This overcomes a design flaw
that would otherwise cause the source cluster mirrors to not come up.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>