Don't print stack traces on errors
Stack traces for errors (rather than panics) are not very helpful and
can overwhelm the user. Therefore, only print stack traces on panics,
and simply return errors.

We also log errors with stack traces to help with debugging. However, we
do this at the debug level so that its not printed twice to
stdout/stderr.

Co-authored-by: David Kimura <dkimura@vmware.com>