revert: use gprecoverseg in copy mode
For performance on large clusters use gprecoverseg instead of rsync when
reverting in copy mode.

Restoring the master and primaries is needed in copy mode due to an
issue in 5X where the source cluster is left in a bad state after
execute. This is because running pg_upgrade on a primary results in a
checkpoint that does not get replicated on the mirror. Thus, when the
mirror is started it panics and a gprecoverseg or rsync is needed.