upgrade: replace "--" with "dd" in the ID string
GPDB has a bug that doesn't handle "--" in directory names in some
cases.  It is easier to fix here.  We are not materially increasing
the change of a collision with an existing ID by doing this.

We keep generating IDs until we obtain one whose string representation
does not contain "--".