migration_scripts: handle name type columns
We need to handle name type columns that are not in the first column.

Our solution runs this in pre-upgrade:

ALTER TABLE table ALTER COLUMN col TYPE varchar(63)

on all non-(child partition) tables for which col is originally of
name type.

Unfortunately, there are documented and undocumented cases where this
ALTER COLUMN fails.  So we print out to the executing log to inform
the user that they must manually update those tables.

For post-revert, we do nothing now as we do not want to track which
tables the user might have manually migrated.

For post-upgrade, we do nothing now as the name type should not be
used in user tables.

Co-authored-by: Bhuvnesh Chaudhary <bchaduhary@vmware.com>