update golang to 1.15
Since golang is backward source compatible, we update to the current
latest version of golang for our gpupgrade compilation as well as
testing.

A test that makes sure the bats tests are run with this version of
golang is also included.  We were previously running a golang version
of 1.13.X, even though we were supposed to be on 1.14, so a test to
catch this in the future is in order.

We created a new docker image for this PR, pivotaldata/gpdb6-centos7-test-golang,
which is a GPDB run environment that contains golang 1.15 pre-installed.