use default port range
This requires AssignDatadirsAndPorts to take in a port range.

The next commit adds the confirmation dialgue to initialize. This
entails printing out the default values from gpupgrade_config. However,
the default port range is empty which does not reflect to the user what
the actual default is. This force a default such that the CLI can
correctly present the default port range to the user. The default port
range is 50432-65535.