Makefile: add enterprise and oss package targets
This creates RPM packages for enterprise and oss releases based off
their tarball equivalents. The main difference being the RELEASE type is
set to either Enterprise or Open Source for the gpupgrade version
command. Also, the open_source_licenses.txt file is included in the
enterprise package. Also, the name in the Summary field for the RPM
differs.

Co-Authored-by: Xin Zhang <zhxin@vmware.com>
Co-authored-by: Kalen Krempely <kkrempely@vmware.com>
Co-authored-by: David Krieger <dkrieger@vmware.com>