initialize: refactor parse_config to allow for gpupgrade_config testing
This refactors the internals of ParseConfig to separate out transforming
the gpupgrade config file name from the corresponding flag name as well
as erroring out on empty values.

This allows us to test the committed gpupgrade_config file, which has
empty values to indicate to the user that they neeed to fill those in
before running gpupgrade initialize.