ci: update gpupgrade rpm references
Since the rpm got renamed update references in the pipeline and CI.

Use "gpupgrade-intermediate.el7.x86_64.rpm" for the intermediate rpm.
This has the added benfit of using the notion "gpupgrade-*.rpm"
consistently everywhere to both grab the intermediate and release
candidate rpm.

Always check for the license file now that it is included in both the
Open Source and Enterprise releases. There is no harm in included the
license for both releases and simplifys the code.