data-migration-scripts: fix paths and tweak names
Update the migration binaries to find the correct script location.
During RPM installation the data-migration binaries are located in
`/usr/local/bin` and the scripts are in
`/usr/local/bin/greenplum/gpupgrade/data-migration-scripts/`.

The RPM renames the scripts to gpupgrade-migration-sql-generator.bash
and gpupgrade-migration-sql-executor.bash. Update the usage and help
accordingly.