hub: ensure agentConns are ready
Centralize the location of ensuring agentConns are ready inside
step.Begin as this ensures the s.agentConns member variable is not nil
and safe to use.

However, having the s.agentConns inside the step framework indicates
heavy tech debt which needs to be addressed. However, for the time being
centrally call it here so that subsequent calls to s.agentConns are
populated and updated.