ci: add IsViewOrRuleDdl to 5X
Currently, both the source and target dump’s filter double precision
values to `XX`. This normalizes all double precision values to enable
the dumps to match. However, the `FormatViewOrRuleDdl` was taking
precedence over the `ReplacePrecision` filter. Specifically, the
`FormatViewOrRuleDdl` filter was merging the create view statement with
the view definition into a single line. This caused the
`ReplacePrecision` filter to not be able to detect double precision
values in views for the extension jobs.

Thus, as a workaround add the `FormatViewOrRuleDdl` to the source filter
such that both the source and target filters perform both
`ReplacePrecision` and `FormatViewOrRuleDdl` filters.

Co-authored-by: Orhan Kislal <okislal@vmware.com>