rsync: refactor gRPC to be simple pass thru
To keep things clear and consistent and easy simply have the gRPC call
for rsync match the actual rsync parameters. This eliminates the need
for RsyncPairs at the cost of minor duplication of Options and
ExcludedFiles. This overhead is very minor compared to the
maintainability and extensibility of the code.