finalize: rework rename directories
Introduce an enum to clearly specify if only the target cluster needs to
be renamed, or if both the source cluster needs to be archived and the
target cluster renamed.

Now that renaming directories has been moved to the end of finalize
rework renaming directories. That is, in link mode the mirrors have been
deleted to save disk space, so only rename the target directories.