CI: Remove orafce tests and fix hstore setup
The orafce extension has been updated in GPDB6 and the versions
are incompatible with each other. This commit removes
its tests since we suggest the users to uninstall/reinstall.

The hstore setup had an error due to a possibly mishandled
rename operation. This commit fixes it to properly establish
dependency.