finalize: refactor update catalog
Remove UpdateCatalogAndClusterConfig() from the server. This allows for
easier testing, and follows the pattern that step functions do not need
to belong on the server object. That bloats the server unnecessarily.

Remove unneeded validation check as we would encounter far greater
problems before this.

Use sql.DB to update the catalog as this follows a clear and consistent
pattern throughout the code base.

Simplify UpdateGpSegmentConfiguration() to loop through the primaries
and mirrors rather than their content ids.

Update the segment configuration using dbid rather than role and content
id.