finalize: refactor add mirrors
Inline fts time out. No need to have it be a global since its used in
one place.

No need ot pass in the state directory as its a global. Use the utils
helper function.

No need to complicate writing to a file, just use os.WriteFile. We can
mock out the state directory, so write the file inside the function.

No need to pass in the mirrors, just pass in the cluster as is and have
the function do the work.

Remove unneeded and overly complicated tests. We only need to test the
interesting logic.