Enable revert operation on initialize for no mirror systems
Current implementation of revert has a hard stop on systems
without mirrors. While this is necessary for production
systems, it hinders the ability to quickly test on dev
environments for customers. This commit adds a new check
to determine if we are at a stage before execute and
allows the revert in such cases.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>