finalize: fix tablespace rename bug in link mode
In link mode finalize would fail with the following error when tablespaces
mirror directories where a subdirectory of the primary such as
/tmp/testfs/gpseg0 and /tmp/testfs/mirror/gpseg0:

Error: Finalize: rpc error: code = Unknown desc = substep "UPGRADE_MIRRORS": 4
errors occurred:
* rpc error: code = Unknown desc = 2 errors occurred:
* rename
/tmp/testfs/mirror/gpseg3/25260/5 /tmp/testfs/gpseg3/25260/14: no such file or
directory
* rename /tmp/testfs/mirror/gpseg3/25259/5
/tmp/testfs/gpseg3/25259/14: no such file or directory

Previously, the code was renaming on the wrong host. Rather the rename needs to
occur on the mirror tablespace host, which for link mode is the intermediate
mirror host.