Fix index related data migration scripts
The data migration scripts have multiple scripts to handle indexes. If a
column is of name or tsquery type and has an index, this index has to be
dropped before the type can be changed. If this column also belongs to a
partitioned table, we generate a second drop index query. Since we
cannot drop the same index twice, we need the code to accomodate for
this possibility.
In addition, a column is the partition key, its type cannot be altered.
This commit adds a table to the test setup and drops it in the unfixable
objects script to remind us that fact.