Fix post-revert scripts for index inheritance
5X and 6X implement indexes on partitioned tables differently. On 6X,
an index on the root table will inherit the indexes in the child tables
if they have the correct name. On 5X this mechanism doesn't exist. Thus,
our post-revert scripts have to reflect.
This commit reverses the order of index creation for post-revert. In
addition, it adds a check to ensure inherited indexes do not get created
a second time. This setup ensures that the user defined indexes on child
tables are preserved post revert.