Enable option ON_ERROR_STOP on psql calls
If a sql file is executed by psql without this option turned on, any
errors in sql execution will be ignored and psql will keep executing sql
statements. The resulting error code will be the status of the last SQL
statement executed. By forcing psql to stop if an error is encountered
we can stop SQL execution and return with a non zero error code.