data-migration-script to drop parent partitions with seg entries
The ICW regression database has more AO and AOCO parent
partitions with seg entries which is currently causing the 5 to 6
upgrade jobs to fail. Thus, the easiest way to address all these tables
seems to add a pre-initialize data migration script. There is no need
for a post-finalize or post-revert script since the seg entries are not
needed.