If the cluster is stopped don't check active connections
That is, if coordinator is not running assume the cluster is stopped and
there are no active connections. This prevents the active connection
check from erroring out and customers needing to start their cluster
simply to check there are no active connections and then the following
substep shut it down.