add initial 6-to-7 data migration seed scripts
This is to support 6-6-cluster jobs which is already running, but never
had proper 6-to-7 seed scripts. Certain SQL was failing which we were
ignoring. This commit allows us to no longer ignore errors and helps us
start to support 6-to-7 data migration scripts. As an initial attempt to
support 6-to-7 seed scripts copy and paste the 5-to-6-seed-scripts and
tweak the necessary scripts. These 6-to-7-seed-scripts are not
production ready.

Specifically:
1) Update initialize/tables_using_name_and_tsquery scripts to work on 6X
by  changing `unnest(attrnums) attnum` to numsegments when querying
gp_distribution_policy.

2) Remove gphdfs_user_roles