introduce parent_backup_dir
The parent directory is the location used internally by gpupgrade to
store the backup of the master data directory and user defined master
tablespaces. It defaults to the parent directory of the master data
directory such as /data. That is, given a master data directory of
/data/master/gpseg-1 the parent-backup-directory will be /data.

The coordinator host backup directory contains the pre & post upgraded
coordinator data directory, and the user defined coordinator
tablespaces. The primary segment hosts backup directory contain the post
upgraded coordinator data directory. Note that the backup directory is
created on all segment hosts for code convenience and simplicity.

Note: We do not display this parameter in the gpupgrade_config as the
default is dependent on the master data directory which is only found
after gpupgrade is run. We clearly document it, and give detailed next
actions to users when an error occurs.