revert: always rsync in link mode
We should rsync source mirrors to source primaries if pg_upgrade ever
runs with link mode. Technically this is when coordinator upgrade is
run. Since this happens early in execute, its cleaner to always call
rsync which should be a no-op if there is no data to transfer.

For large clusters rsync will take some time to calculate diffs even for
a no-op, but we are ok with this trade-off for code simplicity.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>