Update expected extension.diff due to AGGREGATE bugfix
The public.max and public.min aggregates were now dumping out the SORTOP
which added dependency on public.> and public.< citext operators. Due to
that dependency, the public.> and public.< citext operators were moved
upwards from their original location in the schema dumps. Originally,
the SORTOP would not be defined which caused the 5X=>6X upgrade to not
carry over the SORTOP. This was actually a bug that was fixed in the
pg_dump optimization changes because the SORTOP is definitely needed.
The original bug was discovered to be in the pg_dump.c dumpAgg function
where aggsortop was not being queried and parsed correctly. When running
against GPDB 5X, the query would contain
aggsortop::pg_catalog.regoperator which would give an output of
">(citext,citext)" whereas the 6X query with no casting would give back
the operator oid. Later in the dumpAgg processing logic, it'll use that
aggsortop output as an argument to function getFormattedOperatorName()
which expected the operator oid. Because an incorrect argument was being
passed in when querying against 5X with the 5X query, the SORTOP was
never outputted in the SQL dump and never restored to the target 6X
cluster during pg_upgrade. The pg_dump optimization changes fixes this
by never casting the aggsortop. Hence, we only need to do a diff filter
fix.

GPDB6 reference commit with SORTOP fix:
https://github.com/greenplum-db/gpdb/commit/fb2922001068d6aee69f2a80ec805ca54ad1091f

Co-authored-by: Jimmy Yih <jyih@vmware.com>