ci: rename jobs
Rename jobs to follow the format source-to-target-osVersion-name such as
5-to-6-centos7-acceptance-tests.

Its much more clear to have the OS version next to the version name
since the main information one needs when reading is the source, target,
and OS version. This belongs as a clump of information together rather
than separated by the test name.