pg_upgrade: add upgradeable aoseg entries tests
Ensure no pg_aoseg entries are created for AO or AOCO tables when there
are no tuples to add. That is, only create a pg_aoseg entry when the
tupcount is greater than 0. This ensures post-upgrade vacuum freeze
succeeds. Entries can be created when there are no tuples to add when
altering the distribution of an AOCO partition table which does CREATE
and INSERT across, or when inserting an empty row.

Having empty pg_aoseg entries caused pg_upgrade vacuum freeze errors.
For example, while upgrading the primaries a VACUUM FREEZE on a root
AOCO table with a non-empty pg_aocsseg table on the master resulted in:
  "ERROR","58P01","could not open Append-Only segment file
  ""base/16410/20863.1665"": No such file or directory",,,,,,"VACUUM
    (FREEZE);",0,,"aomd.c"