Add test to verify AO tables with unused aoblkdirs can be upgraded
When an index is created for an AO table an aoblkdir relation gets
created for the index. But when all indexes are dropped for the table,
the aoblkdir does not get removed. This scenario needs to be handled
gracefully by gpupgrade. This test ensures The aoblkdirs for AO tables
with no indexes are filtered out and not upgraded to the target cluster.

With the changes to upgrading AO aux tables using relfilenodes, this is
now the expected behavior.

Upgrade AO aux table using relfilenodes reference commit:
https://github.com/greenplum-db/gpdb/commit/466df4e26a1e456f