data migration scripts: refactor global scripts
Global scripts are ones that are only applied once for the postgres
database, rather than multiple times for each database.

Refactor to check if a seed script is global rather than if the seed
script directory is global. The impetus for this change is to combine
the cluster_stats and database_stats into a single directory such that
customers are forced to run "all" stats scripts rather than "some".