Move some unit tests into their respective test packages
General golang testing best practices is to have public functions be
tested in respective test packages instead of the actual
package. There's a handful of gpupgrade unit tests that do not do
this. This patch takes care of the trivial ones. The remaining ones
can be dealt with later as they require some minor discussion and
thought.