bootstrap config.Create with all known values
Add all known values when first creating the config file.

Revert no longer needs to call Config.GetInitializeConfiguration since
the upgradeID and other known values are bootstraped in the config.
Having good knonw values in the config allow the rest of revert to
succeed.

GetInitializeConfiguration no longer needs to save the same parameters
that have already been bootstrapped in the config.

Bootstrap with known values early on so helper functions can be used.
For example, bootstrap with the hub port such that connectToHub can
succeed. Bootstrap with the source and target cluster GPHOME's, and
source cluster port such that when initialize exits early, revert has
enough information to fill in the rest of the config values so that the
rest of revert can succeed.

Note: Bootstrapping these values overloads the hub's persisted
configuration with that of the CLI when ideally these would be separate