Remove data migration progress bars
The progress bars that were being outputted in data migration generate
and apply were not working as we expected. This is because the
progress bar library we were using was limited to one progress bar
that required dedicated stdout usage to do its single progress bar
rendering.  We were using it to output multiple progress bars with
other stdout messages that could be interleaved; hence the incomplete
progress bars showing up sometimes.  We'll remove the progress bars
for now and revisit with a different library that can support our
needs.