Topological sort dependant views using deprecated tsquery
To drop views using deprecated types in the correct order, they
must be sorted based on their dependencies. The previous sort
implementation was coded under the assumption that a view could only
depend on 1 other relation. This resulted in a few unaccounted edge
cases like the ones listed below.

 1. A view depending on 1 view and 1 table
 2. A view depending on 2+ tables
 3. A view depending on 2+ views

Trying to fix these edge cases for the old custom sorting algorithm
uncovered more issues and the code logic quickly became unreadable and
very difficult to follow. Redoing the sorting using a standard
topological sort is the right thing to do as it would guarantee the
correct view order and would be much easier to maintain moving forward.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>