Add non-upgradable test for tables distributed on duplicate columns
Because of a GPDB5 bug in it's parser, it is possible to create tables
distributed on duplicate columns. Ensure pg_upgrade does not proceed
with upgrade and generates a txt file with the problem tables. Since
this is patched on GPDB6 allowing pg_upgrade to continue would result in
pg_restore failure during `gpupgrade execute`.