refactor config Write() and Read()
Instead of json.NewEncoder and json.NewDecoder use json.Marshal and
json.Unmarshal which do not need us to pass in a buffer or create a byte
reader. It simplifies the code. Additionally, json.Encoder and
json.Decoder is geared towards dealing with streaming data versus data
from a writer io interface.