Check more often that the source cluster is synchronized before upgrade
We already check that the cluster is synchronized when creating the
source cluster config file at the beginning of Initialize. However,
the state of the source cluster could have changed any time afterwards
between that early Initialize step and upgrading the master segment in
Execute. If the cluster was out-of-sync and we run Revert after
Execute, there is potential for severe catalog corruption. In
particular, this is more likely to happen to the master and standby
master since there's a GPDB-specific mechanism to bypass the
synchronous WAL replication when the standby master is not
responsive. To prevent this, we add a quick sanity check for cluster
synchronization before the Initialize pg_upgrade checks and more
importantly before doing the actual upgrade operation in Execute.