refactor upgrade ID to be a string
Keep it simple! Since upgrade ID is only ever used as a string there is
no reason not to just have it be that type. It also allows us to
serialize it in config.json as a string rather than a random number
making it easy to correlate and debug on customer systems.