Add --pg-upgrade-jobs flag
The value given to gpupgrade's --pg-upgrade-jobs flag propagates to
pg_upgrade's --jobs flag.  The pg_upgrade --jobs flag parallelizes the
number of databases that it dumps and restores at a time. The default
value for --pg-upgrade-jobs is set to 4. It was an arbitrary number that
was chosen because it seemed like a safe default.