Remove distribution column filter in tsquery ALTER statement generation
Tsquery types cannot be part of a distribution key. When generating
ALTER statements that modify tsquery datatype to text datatype there is
no need to check if the tsquery column is being used as a distribution
key. This was likely leftover code when gen_fix_tsquery_to_text.sql was
copied from the script that handled deprecated name types.