Do not generate revert/finalize ALTERs for tables partitioned on tsquery
Altering a column type when the column is a partition key is not
allowed. For this reason, intialize migration scripts for deprecated
tsquery type does not generate an ALTER statement for tables partitioned
on a tsquery column. Tables that are partitioned on a tsquery column
should also not generate an ALTER statement for revert and finalize
migration scripts.

When recreating indexes on partition tables during revert and finalize,
dropped child partition indexes are also brought back. This is a known
behavior that has not been addressed. Tsquery partition indexes
recreation is not immune to this behavior.

Commit where tsquery ALTER statements were disabled in intialize for tables
partitioned on a tsquery column:
1c455b85e44ce5ff2426ece00ee97fbd77c6e1d1