fix progress bar
Need to wait for "both" the work being done and for the progress bars
being incremented (ie: rendered).

The progress bar is tied to the number of seed scripts executed rather
than the number of actual generated scripts written. This is because the
progress bar needs to know up front how many units of work to do. It
makes sense to increment the bar right after the seed scripts are
executed since this is what the actual bar is tied to. The previous
implementation had that.

However, there is a small section of code in which the bar is
incremented before the generated script is fully written to. This may
show the bar being completed even though there is still a little work to
do in writing the file. In reality this is negligible since for even
extremely large scripts the file is written very fast on current
systems.

Given all of the above some people find it easier to increment the bar
after the script was written even though this is technically not
correct. Given the way the code is structured this results in the bar
being incremented in two places: 1) once after the seed scripts are run
and there is no generated script output to write, and 2) after the seed
scripts are run and after the generated script is written.

So it is a tradeoff of having the increment scattered in two places
versus a negligible scenario where the progress bar could show being
completed while the file is being written.