refactor hub gRPC responses
Initially gpupgrade had the design intention of loosely following the
MVC (model view controller) concept for the CLI and Hub. Having the CLI
be very simple with very little to no business logic, so that eventually
the CLI could be run from any machine rather than just the coordinator.
In practice this is not much of a requirement and one that is not
prioritized. Additionally, most if not all of the other Greenplum
products are required to run from the coordinator.

If one were to keep this intention we could be using protobuf for all
our internal objects like the Cluster object such that the CLI and hub
could have access to all the useful helper functions such as
RunGreenplumCmd(), WaitForClusterToBeReady(), etc. This would avoid
needing to serialize and deserialize between protobuf and internal
golang structs. However, that is quite a big change.

Alternatively, we could break the above intention and read from
gpupgrade’s config.json on the CLI to have access to gpupgrade's state
and objects. This is simple and easy to do.