ci: use future minor version RPMs
Releng cannot produce a release candidate RPM that once tagged produces
a corresponding RPM with the correct version. Rather another commit is
needed to correctly increment the version. This causes lots of headaches
for gpupgrade since we compile in a minimum GPDB version causing
gpupgrade release to wait for the correct version of GPDB to flow
through the pipelines. For now, use a release candidate that has the
future minor version rather than the current minor version. This is
problematic since the commit sha and this future version do not
correctly match reality.

Remove unused majorVersion function