fix pg_upgrade expectations for non-debug builds
Now that we are no longer using a deubug build the following assertions
need two different result files 1) the default result file for optimized
non-debug builds ran on the CI, and 2) a debug specific result file for
local testing. See previous commit where we are using future minor
version rpm for testing which is a non-debug build.