gpupgrade tests move defer revert
Move defer revert to right after initialize in case their is a failure
earlier on. That is, the defer should be right next to the resource that
needs to be cleaned up which in this case is initialize.

One could potentially add a revert inside each of the initalize and
execute functions when an error occurs inside them. This may make
debugging harder.