consolidate acceptance tests: mv gpupgrade tests up one dir
There are two types of acceptance tests, gpupgrade and pg_upgrade. As
both of these tests are ported from bats to golang, we need to move the
respective tests out of their subdirectory to put them in the same
package. These two types of tests use the same set of support functions
(initialize, execute, finalize, revert, restart, kill-services). As more
bats tests are ported to golang we needed to make one fo the following
choices.

1. Keep the gpupgrade and pg_upgrade tests in separate test packages.
   This would result in many end-to-end specific support functions
   landing in testutils and would result to a LOT of testutils.xxxxxx
   calls in both tests. Dot import could have been used to resolve this,
   but dot import is only recommended for use to break circular imports.
   Keeping the support functions in gpupgrade_tests test package and
   importing into pg_upgrade_tests package is not an option since golang
   does not allows functions to be exported from *_test.go files.

2. Consolidate the gpupgrade and pg_upgrade tests into one testing
   package. This would be achieved by moving the contents of both
   test/acceptance/gpupgrade and test/acceptance/pg_upgrade
   up one directory and remove the nested folders.

We chose to to go with #2. Consolidate gpupgrade and pg_upgrade tests.