port finalize.bats to golang
There is an outstanding FIXME when running validate_mirrors_and_standby after
upgrading in link mode with tablespaces. That is, when the
validate_mirrors_and_standby script is run in link mode then the above call to
remove tablespaces fails with:
  ERROR: failed to acquire resources on one or more segments (SQLSTATE 58M01)

Specifically, it errors when accessing objects in a database created in a
tablespace. That is, accessing a table from a database created in a tablespace
when this test is run in link mode. For example, connecting to foodb and doing
\d+ errors with:
  ERROR:  failed to acquire resources on one or more segments DETAIL:  FATAL:  could not read relation mapping file "pg_tblspc/16389/GPDB_6_301908232/16487/pg_filenode.map": Success (relmapper.c:660)

Connecting to one of the segments errors with:
  PGOPTIONS="-c gp_session_role=utility" psql -p 25432 foodb psql: FATAL:  could not read relation mapping file "pg_tblspc/16389/GPDB_6_301908232/16487/pg_filenode.map": Success (relmapper.c:660)

Inspecting the tablespace location is extremely chaotic especially after
upgrading!