Fix failing tablespace test during finalize
When running the test
"in link mode gpupgrade finalize should also delete mirror directories and honors --use-hba-hostnames"
an error would occur when trying to query tables in a user created
tablespace.

The body of the testFinalize function has several deferred functions to
clean up the test environment. These will be executed in last in first out (LIFO)
order.

Currently this means that the tablespace cleanup functions are run in this order:
1. MustDeleteTablespaces
2. MustRemoveAll(t, tablespaceDir)
3. MustExecuteSQL -- Drops table 'should_be_reverted' created in tablespace

Since we remove the tablespaces and tablespace directory first, the SQL
to drop 'should_be_reverted' will fail.

Fix: Remove the calls to MustDeleteTablespaces and MustExecuteSQL.

MustRemoveAll(t, tablespaceDir) takes care of cleaning up the temp
tablespace directory and the defer call to restoreDemoCluster will
return the state of the source cluster to before the test was run
after testFinalize returns or there is a Fatalf called. We don't
need to explicity drop the tablespace.

Because restoreDemoCluster is always called after testFinalize runs,
creating and dropping the 'should_be_reverted' table within
testFinalize isn't necessary.

Authored-by: Brent Doil <bdoil@vmware.com>