Expand gp_upgrade_upgradable_test to exercise finalize
The typical customer workflow is to run initialize, execute, finalize
and run finalize migration scripts. This change add the two steps to run
finalize and finalize migration scripts as part of the test.

In reference to the test output change.
A bitmap of pages can be created dynamically by the planner and is
then discarded at the end of the bitmap index scan (not to be confused
with a BRIN index). Before, the planner was generating this bitmap for
the btree index table_with_name_columns_idx. After rebuilding statistics
the planner chooses a standard Index Scan. This is expected. Since the
finalize rebuilds statistics, the query plan result changed is expected.