Fix heterogeneous partition table test and remove matchsub
In the midst of all the complex merging of multiple large PRs that
included changes in testing infrastructure, directory structure,
filenames, testing patterns, and test porting some mismerged happened. A
patch using matchsubs was put in to fix the
heterogeneous_partition_tables test. While matchsubs is an acceptable
way to make the test pass, the team has decided on to split 5X and 6X
for exact answer file matching in favor of using matchsub. This change
brings the isolation2 answer files in line with the team practices.