7x ci: create dummy gp_toolkit for 7x pg_isolation2_regress running on 6x
As of GPDB7 gp_toolkit is an extension. There were changes to 7x's
pg_isolation2_regress to create this extension when
pg_isolation2_regress is invoked. To prevent this CREATE statement from
causing erroring out, install a dummy gp_toolkit as a workaround.

Fix sample test to be aligned with 7x's pg_isolation2_regress output

7x's pg_isolation2_regress now requires a setup file.