Add 7x support to QueryPgStatActivity looking for active connections
7x pg_stat_activity query uses the client_addr to filter out potential
internal processes.

From the postgres docs:
client_addr | inet | IP address of the client connected to this backend.
If this field is null, it indicates either that the client is connected
via a Unix socket on the server machine or that this is an internal
process such as autovacuum.

source:
https://www.postgresql.org/docs/12/monitoring-stats.html